Separate pod tracker
